+++ src/main/java/techcourse/myblog/domain/Article.java
@@ -4,7 +4,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import techcourse.myblog.exception.AuthException;
+import techcourse.myblog.service.exception.AuthException;
 
 import javax.persistence.*;
 import java.util.ArrayList;
