+++ src/main/java/techcourse/myblog/service/dto/UserDto.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.dto;
+package techcourse.myblog.service.dto;
 
 import lombok.*;
 import techcourse.myblog.domain.User;
