+++ src/main/java/techcourse/myblog/service/exception/ValidUserException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.exception;
+package techcourse.myblog.service.exception;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
