+++ src/main/java/techcourse/myblog/web/controller/ArticleController.java
@@ -6,8 +6,8 @@ import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 import techcourse.myblog.domain.Comment;
-import techcourse.myblog.dto.ArticleDto;
 import techcourse.myblog.service.ArticleService;
+import techcourse.myblog.service.dto.ArticleDto;
 import techcourse.myblog.web.UserSession;
 
 import java.util.List;
