+++ src/main/java/techcourse/myblog/web/interceptor/LoginInterceptor.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.interceptor;
+package techcourse.myblog.web.interceptor;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
