+++ src/main/java/techcourse/myblog/web/support/ExceptionAdvice.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.support;
+package techcourse.myblog.web.support;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -7,7 +7,7 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
-import techcourse.myblog.exception.AuthException;
+import techcourse.myblog.service.exception.AuthException;
 
 @ControllerAdvice
 public class ExceptionAdvice {
