+++ src/main/resources/application.properties
@@ -7,9 +7,10 @@ spring.h2.console.enabled=true
 spring.h2.console.path=/h2-console
 spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
 spring.datasource.initialization-mode=always
-# jpa sql view
+# jpa
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
+spring.jpa.open-in-view=false
 # session
 server.servlet.session.tracking-modes=cookie
 # logging
@@ -18,4 +19,4 @@ spring.output.ansi.enabled=always
 logging.level.techcourse.myblog=DEBUG
 # livereload
 spring.devtools.livereload.enabled=true
-spring.thymeleaf.cache=false    
+spring.thymeleaf.cache=false
