+++ src/test/java/techcourse/myblog/domain/ArticleTest.java
@@ -2,7 +2,7 @@ package techcourse.myblog.domain;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import techcourse.myblog.exception.AuthException;
+import techcourse.myblog.service.exception.AuthException;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
